+++ src/main/web/server/log.py
@@ -6,6 +6,7 @@
 # +=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=+=
 
 
+import conf
 import os
 import sys
 from datetime import datetime
@@ -14,8 +15,7 @@ from typing import Final
 
 ENCODING: Final = "utf8" # Put this here because python can't import this field statically from other files :(
 
-LOG_FILE_SYS_PROPERTY: Final = "pc_server_log_file"
-STDLOG_FILE: str = os.getenv(LOG_FILE_SYS_PROPERTY)
+STDLOG_FILE: Final = conf.SERVER_LOG_FILE
 
 DEBUG: Final = 0
 INFO: Final = 1
@@ -100,10 +100,6 @@ def stdlog(level: int, location: str, message: str, prnt: bool = False) -> None:
         print(f"WARN (log)  stdlog called with invalid arguments, cannot log")
         print(f"WARN (log)  \tmessage: {message}")
         return
-
-    global STDLOG_FILE
-    if (STDLOG_FILE == None and os.getenv(LOG_FILE_SYS_PROPERTY) != None):
-        STDLOG_FILE = os.getenv(LOG_FILE_SYS_PROPERTY)
 
     if (prnt == False):
         prnt = (level == WARN or level == ERROR or level == FATAL)